+++ pyinterpolate/idw/__init__.py
@@ -1,0 +1,1 @@
+from .idw import inverse_distance_weighting